backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
requests
gradio_client
python-jose[cryptography]
passlib[bcrypt]
